+++ taskmaster/Cargo.toml
@@ -6,4 +6,5 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-yaml-rust = "0.4"
+serde = { version = "1.0", features = ["derive"] }
+serde_yaml = "0.9"
