+++ taskmaster/build.rs
@@ -1,0 +1,4 @@
+fn main() {
+    // Tell Cargo that if the given file changes, to rerun this build script.
+    println!("rustc-link-lib=c");
+}
