+++ taskmaster/template.conf
@@ -1,34 +1,49 @@
-programs:
-    nginx:
-        cmd: "/usr/local/bin/nginx -c /etc/nginx/test.conf"
-        numprocs: 1
-        umask: 022
-        workingdir: /tmp
-        autostart: true
-        autorestart: unexpected
-        exitcodes:
-            - 0
-            - 2
-        startretries: 3
-        starttime: 5
-        stopsignal: TERM
-        stoptime: 10
-        stdout: /tmp/nginx.stdout
-        stderr: /tmp/nginx.stderr
-        env:
-            STARTED_BY: taskmaster
-            ANSWER: 42
-    vogsphere:
-        cmd: "/usr/local/bin/vogsphere-worker --no-prefork"
-        numprocs: 8
-        umask: 077
-        workingdir: /tmp
-        autostart: true
-        autorestart: unexpected
-        exitcodes: 0
-        startretries: 3
-        starttime: 5
-        stopsignal: USR1
-        stoptime: 10
-        stdout: /tmp/vgsworker.stdout
-        stderr: /tmp/vgsworker.stderr
+nginx:
+    cmd: "/usr/local/bin/nginx -c /etc/nginx/test.conf"
+    numprocs: 1
+    umask: 022
+    workingdir: /tmp
+    autostart: true
+    autorestart: unexpected
+    exitcodes:
+        - 0
+        - 2
+    startretries: 3
+    starttime: 5
+    stopsignal: TERM
+    stoptime: 10
+    stdout: /tmp/nginx.stdout
+    stderr: /tmp/nginx.stderr
+    env:
+        STARTED_BY: taskmaster
+        ANSWER: "42"
+vogsphere:
+    cmd: "/usr/local/bin/vogsphere-worker --no-prefork"
+    numprocs: 8
+    umask: 077
+    workingdir: /tmp
+    autostart: true
+    autorestart: unexpected
+    exitcodes:
+        - 0
+    startretries: 3
+    starttime: 5
+    stopsignal: USR1
+    stoptime: 10
+    stdout: /tmp/vgsworker.stdout
+    stderr: /tmp/vgsworker.stderr
+ls:
+    cmd: "ls -l -a"
+    numprocs: 8
+    umask: 077
+    workingdir: /tmp
+    autostart: true
+    autorestart: unexpected
+    exitcodes:
+        - 0
+    startretries: 3
+    starttime: 5
+    stopsignal: USR1
+    stoptime: 10
+    stdout: /tmp/vgsworker.stdout
+    stderr: /tmp/vgsworker.stderr
